🔨 : socket 파일 분리
